добавлен интерфейс для сервиса
